api/requirements.txt
flask
python-dotenv
pymongo[srv]
requests
flask-cors